feat(user-service): implement user schemas and refactor UserService methods to use schema types

+++ src/user-management/src/services/user.ts
@@ -1,43 +1,41 @@
 import { UserRepository } from '../repositories/user.js';
-import { ActivityStatus } from '../../generated/prisma/client.js';
-
+import * as SchemaTypes  from '../schemas/user.services.js';
 
 export class UserService {
   constructor(private readonly userRepository: UserRepository) {}
 
-  async createUser(email: string, name: string): Promise<{ id: number }> {
-    return await this.userRepository.create({ email, name });
+  async createUser(input: SchemaTypes.CreateUserSchemaType): Promise<{ id: number }> {
+    return await this.userRepository.create( { email: input.user_email, name: input.user_name });
   }
 
-  async deleteUser(id: number): Promise<void> {
-    await this.userRepository.delete({ id });
+  async deleteUser(schema: SchemaTypes.FindUserByIdSchemaType): Promise<void> {
+    await this.userRepository.delete({ id: schema.user_id });
   }
 
-  async updateUser(id: number, email?: string, name?: string): Promise<void> {
-    if (!email && !name) {
+  async updateUser(input: SchemaTypes.UpdateUserSchemaType): Promise<void> {
+    if (!input.user_email && !input.user_name) {
       throw new Error('At least one field (email or name) must be provided for update.');
     }
-    await this.userRepository.update({ id, email, name });
+    await this.userRepository.update({ id: input.user_id, email: input.user_email, name: input.user_name });
   }
 
-  async updateStatus(id: number, activity_status: ActivityStatus): Promise<void> {
-    await this.userRepository.updateStatus({ id, activity_status });
+  async updateStatus(input: SchemaTypes.UpdateUserStatusSchemaType): Promise<void> {
+    await this.userRepository.updateStatus({ id: input.user_id, activity_status: input.activity_status });
   }
 
-  async findUserById(id: number) {
-    return await this.userRepository.findById(id);
+  async findUserById(schema: SchemaTypes.FindUserByIdSchemaType) {
+    return await this.userRepository.findById(schema.user_id);
   }
 
-  async findUserByEmail(email: string) {
-    return await this.userRepository.findByEmail(email);
+  async findUserByEmail(schema: SchemaTypes.FindUserByEmailSchemaType) {
+    return await this.userRepository.findByEmail(schema.user_email);
   }
 
-  async findUserByName(name: string) {
-    return await this.userRepository.findByName(name);
+  async findUserByName(schema: SchemaTypes.FindUserByNameSchemaType) {
+    return await this.userRepository.findByName(schema.user_name);
   }
 
-  async findUsersByStatus(activity_status?: ActivityStatus) {
-    return await this.userRepository.findByStatus({ activity_status });
+  async findUsersByStatus(schema: SchemaTypes.FindUsersByStatusSchemaType) {
+    return await this.userRepository.findByStatus({ activity_status: schema.activity_status });
   }
-
 }
